Randomize image data
Currently, the images generated are not randomized between threads as
the numpy seed is identical for all subprocesses. Replacing the
`RandomState` with `random.seed` updates the application-level numpy
seed without needing to use a returned object, and incrementing the
seed based on the process counter ensures each image is truly random.

Signed-Off-By: Robert Clark <roclark@nvidia.com>