Chunk image creation process
The starmap method from the multiprocessing library allocates memory for
an entire iterable. When using starmap with very long iterables, as can
be the case when creating a large amount of images using Imageinary,
memory-constrained systems and containers can have stalled processes
which leads to failure to complete the workload. By chunking the starmap
into smaller pieces, the application trades slightly more overhead and
slower processing for more stable functionality on a wider array of
systems.

Signed-Off-By: Robert Clark <roclark@nvidia.com>