Make size flag a boolean
The --size flag was being treated as a generic object which required
any value to be passed to be considered True. This meant users had to
specify `--size 1` or similar in order to get the flag to work. Forcing
the value to be True when passed and False otherwise is the original
intended behavior.

Signed-Off-By: Robert Clark <roclark@nvidia.com>